rename and move modal into comp
- removed SwitchtoWethButton
- removed EthFlow from Swap index
