+++ src/custom/pages/Claim/InvestmentFlow/InvestOption.tsx
@@ -33,7 +33,7 @@ import tryParseCurrencyAmount from 'lib/utils/tryParseCurrencyAmount'
 import { calculateInvestmentAmounts, calculatePercentage } from 'state/claim/hooks/utils'
 import { AMOUNT_PRECISION, PERCENTAGE_PRECISION } from 'constants/index'
 import { useGasPrices } from 'state/gas/hooks'
-import { AVG_APPROVE_COST_GWEI } from 'components/swap/EthWethWrap/helpers'
+import { AVG_APPROVE_COST_GWEI } from 'components/swap/EthFlow/helpers'
 import { EnhancedUserClaimData } from '../types'
 import { OperationType } from 'components/TransactionConfirmationModal'
 import { ONE_HUNDRED_PERCENT } from 'constants/misc'
